requirements.txt
gradio>=4.0
matplotlib>=3.7
openai
python-dotenv
